refactor: SearchBar 컴포넌트를 공통으로 사용할 수 있도록 handleSearch props로 전달 받게 구현
